* JPEG 2000, HEIF and JPEG XL read and write what other implementations do (#314)
* + JPEG XL writes a real lossless modular codestream and is registered as a writer

* * the HEIF writer coded PCM units libde265 refuses

H.265 7.4.3.2 caps Log2MaxIpcmCbSizeY at Min(CtbLog2SizeY, 5), so the
64 by 64 PCM coding units were out of range. libde265 rejected every
file with 'coded parameter out of range'; ffmpeg accepted them, which
is why reading our own output back had never shown it. The coding tree
block is now 32 by 32 and each tree block stays one unsplit PCM unit.

The level was also a constant 4.0 regardless of picture size, and the
hvcC record left the constraint flags zero where the sequence parameter
set spells them out. Both now follow the picture.

* + HEIF reading of ten- and twelve-bit intra pictures

libheif and x265 write ten bits unless told otherwise and ImageMagick
writes twelve, so an eight-bit-only reader refused most HEIC files in
existence. The decoded picture now holds its samples in the sequence's
own depth instead of a byte; the transform, dequantiser, intra
prediction and both loop filters were already parameterised by it, and
the range-extension tools that would need more are still refused by
name in the parameter sets.

Checked sample for sample against ffmpeg's HEVC decoder over 49 streams
covering coding tree sizes 16 to 64, quantisers 0 to 51, multiple
slices, transform skip, sign hiding and both loop filters on and off.

* * JPEG 2000 reads and writes what other implementations do

The format table blamed both halves on missing tag trees. Tag trees were not
the fault: they were implemented and wired into both Tier-2 sides already.
Measuring against OpenJPEG rather than against ourselves found the real
defects, all of them the kind a reader and a writer that share them agree
about perfectly.

The MQ coder had the two halves of the coding interval the wrong way round.
T.800 puts the less probable symbol at the bottom, so coding an MPS adds Qe to
the code register; this added the remaining interval instead and decoded with
the matching comparison. That is a different arithmetic code, not a variant of
the same one, and nothing else could follow either side of it.

Tier-1 had four faults of its own. It scanned in raster order where Annex D
scans stripes of four rows and then columns. It never marked a coefficient
coded in the significance-propagation pass, so the cleanup pass of the same
bit-plane coded it again. It used the LL context table for all four subband
orientations, where HL is that table with the horizontal and vertical
neighbour counts exchanged and HH has one of its own. And it took the number
of magnitude bit-planes from the coding-pass count instead of from Mb less the
signalled zero bit-planes, which agrees only for code-blocks coded to
completion — that is, never for a lossy stream.

The wavelet filtered rows before columns going one way and columns before rows
going the other, so each undid the other but neither matched F.3, and 5/3
rounding does not commute. It also took every subband to begin at an even
coordinate, which is what decides which half of a line is low-pass.

Above that the packet layer knew nothing of precincts, of tiles past the
first, of the four progression orders that are not LRCP, or of more than one
quality layer; and the reader picked between two tile-data formats by
comparing lengths, one of them this library's own invention of raw int32
coefficients. That format is gone, so the writer no longer emits a file only
this library can read.

Verified against OpenJPEG 2.5.4, ImageMagick 7 and ffmpeg:

  - 90 codestreams from opj_compress — default options, JP2 container, one to
    five decomposition levels, 16 to 64 sample code-blocks, explicit precinct
    partitions, 37x29 to 64x64 tiles, all five progression orders, tile-part
    splitting, SOP and EPH markers, up to six quality layers, the colour
    transform on and off, vertically causal contexts, segmentation symbols, a
    non-zero image origin, 8- and 16-bit samples — decode sample for sample
    identically to opj_decompress. Lossless is exact; the irreversible 9/7
    files are within 2 of 255 of OpenJPEG's own decode.

  - 48 codestreams from this writer — six pictures at zero, one, two and five
    decomposition levels, raw and in a JP2 container — are read by
    opj_decompress, magick and ffmpeg with zero differing samples.

  - On identical input the output is 39 bytes smaller than opj_compress's,
    that being its COM marker; the coded data is the same size.

Written from ITU-T T.800 Annexes B through F (rung 3). OpenJPEG is licence
compatible, but it was used as the oracle rather than as a source: no code was
taken from it, and every constant comes from the standard's own tables.

* * the format table states what the three codecs now do

* * the reference back-link points at the package readme

+++ Hawkynt.FileFormats.Video/Codecs/H265/H265Deblocking.cs
@@ -224,7 +224,7 @@ internal static class H265Deblocking {
     }
   }
 
-  private static bool _StrongDecision(byte[] plane, int at, int step, int curvature, int beta, int clip) {
+  private static bool _StrongDecision(ushort[] plane, int at, int step, int curvature, int beta, int clip) {
     if (curvature >= beta >> 2)
       return false;
     var span = Math.Abs(plane[at - 4 * step] - plane[at - step]) + Math.Abs(plane[at] - plane[at + 3 * step]);
@@ -234,7 +234,7 @@ internal static class H265Deblocking {
   }
 
   private static void _FilterStrong(
-    byte[] plane, int at, int step, int clip, int maximum, bool keepP, bool keepQ) {
+    ushort[] plane, int at, int step, int clip, int maximum, bool keepP, bool keepQ) {
     var p3 = plane[at - 4 * step];
     var p2 = plane[at - 3 * step];
     var p1 = plane[at - 2 * step];
@@ -259,7 +259,7 @@ internal static class H265Deblocking {
   }
 
   private static void _FilterWeak(
-    byte[] plane, int at, int step, int clip, int maximum, bool filterP1, bool filterQ1,
+    ushort[] plane, int at, int step, int clip, int maximum, bool filterP1, bool filterQ1,
     bool keepP, bool keepQ) {
     var p2 = plane[at - 3 * step];
     var p1 = plane[at - 2 * step];
@@ -274,22 +274,22 @@ internal static class H265Deblocking {
 
     delta = Math.Clamp(delta, -clip, clip);
     if (!keepP) {
-      plane[at - step] = (byte)Math.Clamp(p0 + delta, 0, maximum);
+      plane[at - step] = (ushort)Math.Clamp(p0 + delta, 0, maximum);
       if (filterP1) {
         var adjust = Math.Clamp((((p2 + p0 + 1) >> 1) - p1 + delta) >> 1, -(clip >> 1), clip >> 1);
-        plane[at - 2 * step] = (byte)Math.Clamp(p1 + adjust, 0, maximum);
+        plane[at - 2 * step] = (ushort)Math.Clamp(p1 + adjust, 0, maximum);
       }
     }
 
     if (keepQ)
       return;
 
-    plane[at] = (byte)Math.Clamp(q0 - delta, 0, maximum);
+    plane[at] = (ushort)Math.Clamp(q0 - delta, 0, maximum);
     if (!filterQ1)
       return;
 
     var adjustQ = Math.Clamp((((q2 + q0 + 1) >> 1) - q1 - delta) >> 1, -(clip >> 1), clip >> 1);
-    plane[at + step] = (byte)Math.Clamp(q1 + adjustQ, 0, maximum);
+    plane[at + step] = (ushort)Math.Clamp(q1 + adjustQ, 0, maximum);
   }
 
   private static void _FilterChromaSegment(
@@ -330,9 +330,9 @@ internal static class H265Deblocking {
       var delta = Math.Clamp((((q0 - p0) << 2) + p1 - q1 + 4) >> 3, -clip, clip);
 
       if (!keepP)
-        plane[at - step] = (byte)Math.Clamp(p0 + delta, 0, maximum);
+        plane[at - step] = (ushort)Math.Clamp(p0 + delta, 0, maximum);
       if (!keepQ)
-        plane[at] = (byte)Math.Clamp(q0 - delta, 0, maximum);
+        plane[at] = (ushort)Math.Clamp(q0 - delta, 0, maximum);
     }
   }
 
@@ -340,6 +340,6 @@ internal static class H265Deblocking {
     => frame.IsTransquantBypassAt(block)
        || (sps.PcmLoopFilterDisabled && frame.IsPulseCodeModulatedAt(block));
 
-  private static byte _Move(int from, int to, int limit, int maximum)
-    => (byte)Math.Clamp(Math.Clamp(to, from - limit, from + limit), 0, maximum);
+  private static ushort _Move(int from, int to, int limit, int maximum)
+    => (ushort)Math.Clamp(Math.Clamp(to, from - limit, from + limit), 0, maximum);
 }
